chore: gradle project name + moved some files
