+++ gradle.properties
@@ -1,1 +1,1 @@
-main=it.unibo.samplejavafx.App
+main=it.unibo.wildenc.mvc.view.impl.TempMain
