+++ settings.gradle.kts
@@ -3,4 +3,4 @@ plugins {
     id("org.gradle.toolchains.foojay-resolver-convention") version "1.0.0"
 }
 
-rootProject.name = "sample-javafx-project"
+rootProject.name = "wildenc"
