+++ src/main/java/it/unibo/wildenc/Launch.java
@@ -1,8 +1,8 @@
-package it.unibo.wildenc.mvc.view.impl;
+package it.unibo.wildenc;
 
 import javafx.application.Application;
 
-public class TempMain {
+public class Launch {
     public static void main(String[] args) {
         Application.launch(EntryPoint.class, args);
     }
