+++ src/main/java/it/unibo/wildenc/mvc/view/impl/SimpleCircleRender.java
@@ -1,3 +1,4 @@
+package it.unibo.wildenc.mvc.view.impl;
 /*
 package it.unibo.wildenc.mvc.view;
